robot_detection/requirements.txt
actionlib==1.12.1
addict==2.4.0
albumentations==1.3.0
angles==1.9.12
astor==0.8.1
asynctest==0.13.0
attrs==22.1.0
base_local_planner==1.16.7
bondpy==1.8.5
camera_calibration==1.15.0
camera_calibration_parsers==1.11.13
catkin==0.7.29
certifi @ file:///croot/certifi_1665076670883/work/certifi
charset-normalizer==2.1.1
codecov==2.1.12
controller_manager==0.18.4
controller_manager_msgs==0.18.4
coverage==6.5.0
cv_bridge==1.13.0
cycler==0.11.0
Cython==0.29.32
decorator==4.4.2
diagnostic_analysis==1.9.7
diagnostic_common_diagnostics==1.9.7
diagnostic_updater==1.9.7
docopt==0.6.2
dynamic_reconfigure==1.6.4
exceptiongroup==1.0.0
flake8==5.0.4
fonttools==4.38.0
gast==0.3.3
gazebo_plugins==2.8.7
gazebo_ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.16
idna==3.4
image_geometry==1.13.0
imagecorruptions==1.1.2
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==4.2.0
iniconfig==1.1.1
interactive-markers==1.11.5
isort==5.10.1
joblib==1.2.0
joint_state_publisher==1.12.15
joint_state_publisher_gui==1.12.15
kdl_parser_py==1.13.1
kiwisolver==1.4.4
kwarray==0.6.4
laser_geometry==1.6.7
lmdb==1.3.0
matplotlib==3.5.3
mccabe==0.7.0
message_filters==1.14.13
mmcv==0.2.16
-e git+https://github.com/WXinlong/SOLO.git@95f3732d5fbb0d7c7044c7dd074f439d48a72ce5#egg=mmdet
networkx==2.6.3
numpy==1.21.6
opencv-python==4.5.2.54
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
paddle-bfloat==0.1.7
paddlepaddle-gpu==2.0.1
Pillow==6.2.2
pluggy==1.0.0
pqi==2.0.6
protobuf==3.19.6
pyclipper==1.3.0.post3
pycocotools @ git+https://github.com/cocodataset/cocoapi.git@8c9bcc3cf640524c4c20a9c40e89cb6a2f2fa0e9#subdirectory=PythonAPI
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
pytest==7.2.0
pytest-cov==4.0.0
pytest-runner==6.0.0
python-dateutil==2.8.2
python_qt_binding==0.4.4
PyWavelets==1.3.0
PyYAML==6.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
qudida==0.0.4
requests==2.28.1
resource_retriever==1.12.7
rosbag==1.14.13
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.13
roslaunch==1.14.13
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.13
rosmake==1.14.9
rosmaster==1.14.13
rosmsg==1.14.13
rosnode==1.14.13
rosparam==1.14.13
rospy==1.14.13
rosserial_arduino==0.8.0
rosserial_client==0.8.0
rosserial_python==0.8.0
rosservice==1.14.13
rostest==1.14.13
rostopic==1.14.13
rosunit==1.14.9
roswtf==1.14.13
rqt-moveit==0.5.10
rqt-reconfigure==0.5.4
rqt-robot-monitor==0.5.13
rqt-rviz==0.7.0
rqt_action==0.4.9
rqt_bag==0.5.1
rqt_bag_plugins==0.5.1
rqt_console==0.4.9
rqt_dep==0.4.9
rqt_graph==0.4.11
rqt_gui==0.5.2
rqt_gui_py==0.5.2
rqt_image_view==0.4.16
rqt_launch==0.4.8
rqt_logger_level==0.4.8
rqt_msg==0.4.8
rqt_nav_view==0.5.7
rqt_plot==0.4.13
rqt_pose_view==0.5.8
rqt_publisher==0.4.8
rqt_py_common==0.5.2
rqt_py_console==0.4.8
rqt_robot_dashboard==0.5.7
rqt_robot_steering==0.5.10
rqt_runtime_monitor==0.5.7
rqt_service_caller==0.4.8
rqt_shell==0.4.9
rqt_srv==0.4.8
rqt_tf_tree==0.6.0
rqt_top==0.4.8
rqt_topic==0.4.11
rqt_web==0.4.8
rviz==1.13.24
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
sensor-msgs==1.12.8
Shapely==1.8.5.post1
six==1.16.0
smach==2.0.1
smach_ros==2.0.1
smclib==1.8.5
terminaltables==3.1.10
tf==1.12.1
tf2_geometry_msgs==0.6.5
tf2_kdl==0.6.5
tf2_py==0.6.5
tf2_ros==0.6.5
tf_conversions==1.12.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tomli==2.0.1
topic_tools==1.14.13
torch==1.4.0
torchvision==0.5.0
typing_extensions==4.4.0
ubelt==1.2.2
urdfdom-py==0.4.6
urllib3==1.26.12
xacro==1.13.17
xdoctest==1.1.0
yapf==0.32.0
zipp==3.10.0
